+++ src/main/java/FitMate/FitMateBackend/cjjsWorking/dto/myfit/myWorkout/MyWorkoutReadAllResponse.java
@@ -4,8 +4,7 @@ import FitMate.FitMateBackend.cjjsWorking.service.storageService.S3FileService;
 import FitMate.FitMateBackend.consts.ServiceConst;
 import FitMate.FitMateBackend.domain.BodyPart;
 import FitMate.FitMateBackend.domain.Machine;
-import FitMate.FitMateBackend.domain.Workout;
-import FitMate.FitMateBackend.domain.myfit.MyFit;
+import FitMate.FitMateBackend.workout.entity.Workout;
 import FitMate.FitMateBackend.domain.myfit.MyWorkout;
 import lombok.AllArgsConstructor;
 import lombok.Data;
