+++ src/main/java/FitMate/FitMateBackend/cjjsWorking/service/MyFitService.java
@@ -8,8 +8,8 @@ import FitMate.FitMateBackend.cjjsWorking.exception.errorcodes.CustomErrorCode;
 import FitMate.FitMateBackend.cjjsWorking.exception.exceptions.CustomException;
 import FitMate.FitMateBackend.cjjsWorking.repository.MyFitRepository;
 import FitMate.FitMateBackend.cjjsWorking.repository.RoutineRepository;
-import FitMate.FitMateBackend.cjjsWorking.repository.WorkoutRepository;
-import FitMate.FitMateBackend.domain.Workout;
+import FitMate.FitMateBackend.workout.repository.WorkoutRepository;
+import FitMate.FitMateBackend.workout.entity.Workout;
 import FitMate.FitMateBackend.domain.myfit.MyFit;
 import FitMate.FitMateBackend.domain.myfit.MySupplement;
 import FitMate.FitMateBackend.domain.myfit.MyWorkout;
